requirements.txt
watchdog
docx2pdf
PyPDF2
pywin32
python-dotenv